+++ src/compas_timber/connections/joint_candidate.py
@@ -59,8 +59,8 @@ class JointCandidate(Joint):
 
     def restore_beams_from_keys(self, model):
         """After de-serialization, restores references to elements saved in the model."""
-        self.element_a = model.element_by_guid(self.element_a_guid)
-        self.element_b = model.element_by_guid(self.element_b_guid)
+        self.element_a = model[self.element_a_guid]
+        self.element_b = model[self.element_b_guid]
 
     def add_features(self):
         """This joint does not add any features."""
@@ -90,3 +90,6 @@ class PlateJointCandidate(PlateJoint, JointCandidate):
 
     def __init__(self, plate_a=None, plate_b=None, **kwargs):
         super(PlateJointCandidate, self).__init__(plate_a=plate_a, plate_b=plate_b, **kwargs)
+
+    def _set_edge_planes(self):
+        pass