+++ src/compas_timber/design/workflow.py
@@ -29,10 +29,10 @@ class CollectionDef(object):
 
 
 class ContainerDefinition(object):
-    """Holds a pair of slab and its configuration set if available."""
+    """Holds a pair of panel and its configuration set if available."""
 
-    def __init__(self, slab, config_set=None):
-        self.slab = slab
+    def __init__(self, panel, config_set=None):
+        self.panel = panel
         self.config_set = config_set