+++ src/compas_timber/fabrication/free_contour.py
@@ -1,30 +1,26 @@
-try:
-    from typing import TYPE_CHECKING
+from __future__ import annotations
 
-    if TYPE_CHECKING:
-        from compas.geometry import Polyline  # noqa: F401
+from typing import TYPE_CHECKING
+from typing import Optional
 
-        from compas_timber.elements import Plate  # noqa: F401
-except ImportError:
-    pass
+if TYPE_CHECKING:
+    from compas.geometry import Polyline  # noqa: F401
+
+    from compas_timber.elements import Beam  # noqa: F401
+    from compas_timber.elements import Plate  # noqa: F401
+    from compas_timber.elements import TimberElement  # noqa: F401
 
 import math
 
-from compas.geometry import Brep
 from compas.geometry import Frame
-from compas.geometry import NurbsCurve
 from compas.geometry import Plane
-from compas.geometry import Polyline
 from compas.geometry import Transformation
 from compas.geometry import Vector
 from compas.geometry import angle_vectors
 from compas.geometry import angle_vectors_signed
 from compas.geometry import distance_point_plane
-from compas.geometry import intersection_line_line
-from compas.geometry import offset_polyline
 from compas.tolerance import TOL
 
-from compas_timber.utils import correct_polyline_direction
 from compas_timber.utils import is_polyline_clockwise
 
 from .btlx import AlignmentType
@@ -81,7 +77,16 @@ class FreeContour(BTLxProcessing):
     ########################################################################
 
     @classmethod
-    def from_polyline_and_element(cls, polyline, element, depth=None, interior=False, tool_position=None, ref_side_index=None, **kwargs):
+    def from_polyline_and_element(
+        cls,
+        polyline: Polyline,
+        element: TimberElement,
+        depth: Optional[float] = None,
+        interior: Optional[bool] = False,
+        tool_position: Optional[str] = None,
+        ref_side_index: Optional[int] = None,
+        **kwargs,
+    ):
         """Construct a Contour processing from a polyline and element.
 
         Parameters
@@ -102,17 +107,17 @@ class FreeContour(BTLxProcessing):
 
         if ref_side_index is None:
             ref_side_index = cls.get_ref_face_index(polyline, element)
-
         ref_side = element.ref_sides[ref_side_index]
         tool_position = cls.parse_tool_position(polyline, ref_side, interior, tool_position)
-        depth = depth or element.width
-        transformed_polyline = polyline.transformed(Transformation.from_frame_to_frame(ref_side, Frame.worldXY()))
+        # get_dimensions_relative_to_side [0] returns element dimension normal to ref_side
+        depth = depth or element.get_dimensions_relative_to_side(ref_side_index)[0]
+        transformed_polyline = polyline.transformed(Transformation.from_frame(ref_side).inverse())
         contour = Contour(transformed_polyline, depth=depth, inclination=[0.0])
         return cls(contour, tool_position=tool_position, counter_sink=interior, ref_side_index=ref_side_index, **kwargs)
 
     @classmethod
     def from_top_bottom_and_elements(cls, top_polyline, bottom_polyline, element, interior=False, tool_position=None, ref_side_index=None, **kwargs):
-        # type: (Polyline, Polyline, Plate, bool, str | None, int | None) -> FreeContour
+        # type: (Polyline, Polyline, Union[Plate, Beam], bool, str | None, int | None, dict) -> FreeContour
         """Construct a Contour processing from a list of polylines and element.
 
         Parameters
@@ -135,23 +140,23 @@ class FreeContour(BTLxProcessing):
             ref_side_index = cls.get_ref_face_index(top_polyline, element)
 
         ref_side = element.ref_sides[ref_side_index]
-        xform_to_part_coords = Transformation.from_frame_to_frame(ref_side, Frame.worldXY())
         tool_position = cls.parse_tool_position(top_polyline, ref_side, interior, tool_position)
+        transformation_to_local = Transformation.from_frame(ref_side).inverse()
 
+        top_polyline = top_polyline.transformed(transformation_to_local)
+        bottom_polyline = bottom_polyline.transformed(transformation_to_local)
         if not cls.are_all_segments_parallel(top_polyline, bottom_polyline):  # use DualContour
-            points_principal = [pt.transformed(xform_to_part_coords) for pt in top_polyline]
-            points_associated = [pt.transformed(xform_to_part_coords) for pt in bottom_polyline]
-            contour = DualContour(points_principal, points_associated)
+            contour = DualContour(top_polyline, bottom_polyline)
         else:  # use Contour with inclination
             inclinations = []
             for top_line, bottom_line in zip(top_polyline.lines, bottom_polyline.lines):
                 cp = bottom_line.closest_point(top_line.start)
-                inclinations.append(round(angle_vectors_signed(Vector.from_start_end(top_line.start, cp), -ref_side.normal, -top_line.direction, deg=True), 6))
+                inclinations.append(round(angle_vectors_signed(Vector.from_start_end(top_line.start, cp), Vector(0, 0, -1), -top_line.direction, deg=True), 6))
             if len(set(inclinations)) == 1:
                 inclinations = [inclinations[0]]  # all inclinations are the same, set one global inclination for FreeContour processing
-            depth = distance_point_plane(bottom_polyline[0], Plane.from_frame(ref_side))
-            polyline = top_polyline.transformed(xform_to_part_coords)
-            contour = Contour(polyline, depth=depth, inclination=inclinations)
+            # bottom polyline is parallel to ref_side. Use Z value of first point to calculate depth
+            depth = -bottom_polyline[0][2]
+            contour = Contour(top_polyline, depth=depth, inclination=inclinations)
 
         return cls(contour, counter_sink=interior, tool_position=tool_position, ref_side_index=ref_side_index, **kwargs)  # type: ignore
 
@@ -200,13 +205,14 @@ class FreeContour(BTLxProcessing):
 
     @staticmethod
     def get_ref_face_index(contour_points, element):
-        # type: (Polyline, Plate) -> int
+        # type: (Polyline, Union[Plate, Beam]) -> int
         curve_frame = Frame.from_points(contour_points[0], contour_points[1], contour_points[-2])
         for i, ref_side in enumerate(element.ref_sides):
-            if TOL.is_zero(distance_point_plane(contour_points[0], Plane.from_frame(ref_side)), tol=1e-6) and TOL.is_zero(
-                angle_vectors(ref_side.normal, curve_frame.zaxis, deg=True) % 180.0, 1e-6
-            ):
-                return i
+            distance = distance_point_plane(contour_points[0], Plane.from_frame(ref_side))
+            if TOL.is_zero(distance, tol=1e-6):
+                angle = angle_vectors(ref_side.normal, curve_frame.zaxis, deg=True)
+                if TOL.is_zero(angle, 1e-3) or TOL.is_zero(angle - 180.0, 1e-3):
+                    return i
         raise ValueError("The contour does not lay on one of the reference sides of the element.")
 
     @staticmethod
@@ -233,19 +239,11 @@ class FreeContour(BTLxProcessing):
         :class:`compas.geometry.Brep`
             The resulting geometry after processing.
         """
-        # TODO: this doesn't work anymore with the new definition of FreeContour, needs fixing but currently not sure how to do it.
-        # TODO: this is only called when there features present other than the Plate's outline (i.e. inner cuts)
-        # TODO: should have a look at this also in regards to the global to local transformation of the features
-        ref_side = element.ref_sides[self.ref_side_index]
-        xform = Transformation.from_frame_to_frame(Frame.worldXY(), ref_side)
-        pts = [pt.transformed(xform) for pt in self.contour_param_object.polyline]
-        pts = correct_polyline_direction(pts, -ref_side.normal, clockwise=True)
-        pln = Polyline(offset_polyline(Polyline(pts), 0.001, normal=-ref_side.normal))  # This is the only way I could get the boolean difference to work
-        pt = intersection_line_line(pln.lines[0], pln.lines[-1])
-        pln[0] = pt[0]
-        pln[-1] = pt[0]
-        vol = Brep.from_extrusion(NurbsCurve.from_points(pln, degree=1), -ref_side.normal * self.contour_param_object.depth)
 
+        vol = self.contour_param_object.to_brep()
+        # contour is defined in the ref_side local frame, need to transform first to global then to element local, where geometry is created
+        transformation_ref_side_to_local = element.modeltransformation.inverse() * Transformation.from_frame(element.ref_sides[self.ref_side_index])
+        vol.transform(transformation_ref_side_to_local)
         if self.counter_sink:  # contour should remove material inside of the contour
             return geometry - vol
         else: